requirements.txt
# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Validation
pydantic>=2.0.0

# Security
cryptography>=41.0.0

# EDA & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0

# Logging
loguru>=0.7.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0